requirements.txt
python-telegram-bot==22.5
requests==2.32.5
sympy==1.12
statistics==1.0